[CBRD-21220] index page: move HA to admin column; remove Shard (#5)
http://jira.cubrid.org/browse/CBRD-21220

Removed HA column from quick references. HA entry was moved to admin, Shard entry (korean only) was removed (shard is not recommended, right?)

I've kept the table which will have to be scrolled to the right when it does not fit the screen (all mobile, most laptops).

If the table is split to have fewer columns or if it is replaced with list of list, the table of contents get pushed down and does not make it to the first screen. I think the current design is preferable.

Misc change, mark unlimited width for manual content as important. it will not be overwritten.